fix handling of sreg extension under 2.0
947793fc94341c9d3889e0fa584a1605a85da0b7 looks wrong

The problem is that the version number at the end of the extension URI
will in general have NOTHING to do with the OpenID version number.
The original patch fails if, under openid 2.0,
relying party tries to use ANY extension other than SREG
whose OWN version number is less than 2,
e.g., Provider Authentication Policy Extension (PAPE)
(url = http://specs.openid.net/extensions/pape/1.0),
which will then be rejected because it's not SREG

This backs out that change and instead ensures that the 'sreg' alias is
always used for the SREG extension even under OpenID 2.0,
which appears to be what the original patch was trying to do,
which shouldn't actually matter but will actually help in cases
of b0rken SREG implementations that depend on the alias being 'sreg'.