fix minimum_version confusion
The problem:
    ->args(parameter_object) initializes ->{message}
    to a freshly built IndirectMessage OR undef according as
    the minimum_version check succeeds or fails.  However, this
    assumes ->{minimum_version} has already been initialized;
    if not, then it's 1, the check always succeeds, and anything that's
    supposed to be fail later on because client wants to ignore
    version 1 messages won't, no matter what we subsequently do
    to minimum_version.

    And even if we reorder the initializations so that minimum_version
    is initialized first, we'll STILL fail if minimum_version is later
    changed -- we don't require the client to set everything in the
    ->new() call, so any order remains possible -- and if we
    initialize ->{args} with minimum_version=2 and THEN want
    minimum_version=1, ->{message} has to come back from the
    dead, which it can't because we no longer have the parameter
    object that it was created from.

Bottom line is if the parameters form a viable OpenID message for ANY
protocol version, we have to create an IndirectMessage AND hold onto it.

It's also not clear why we should need to undef ->{args}, the supposedly
low-level parameter access routine, when ->{minimum_version} is raised.
If we change the contract of ->{args} to be the current behavior,
i.e., that the raw getter is ALWAYS there as long as we have a viable
OpenID message for SOME protocol level, then

(1) this won't break any client code that was relying on it being
    there

(2) it won't change Consumer behavior since for all of the places
    where ->{args} is used in Consumer the minimum_version check has
    already occurred, and

(3) this will decouple ->{args} from ->{minimum_version} and we will
    no longer have to care about the ordering of the args() and
    minimum_version() calls.

which seems like an obvious win.

So,
(*) do NOT pass minimum_version to IndirectMessage->new
(*) put a minimum_version check into message() and make
(*) sure everything that looks at {message} is doing so via the accessor